// src/Types/enum.ts
export const enum Color {
   Brown = '#e94340',
   BlurBrown = '#f06462',
   Blue = '#15d2e7',
   Green = '#4ea952',
   Purple = '#952daf',
   Orange = '#fc960e',
   BlurOrange = '#ffb95e96',
   Silver = '#585858',
   MediumSilver = '#a8a7a7',
   LiteSilver = '#d8d8d8',
   White = '#FFF',
   Black = '#000'
}

export const enum ColorHover {
   Brown = '#fc5151',
   Blue = '#1ae4fa',
   Green = '#66f06d',
   Purple = '#dd55ff',
   Orange = '#f1a239',
}